Added a contract
The main concern of the contract is that it is a single test suite
that can be run in multiple implementations of the quote api, the system
under test.

Moved a first draft implementation of finding quote into production, as
well as its internal model of loan amount.

Tests are added regarding invalid given loan amounts, so that the api
responds with a problem.