Added the driven application interface
Since we do not know how and where the list of lenders is supplied; we
have opted to define it as different concern. A concern that is
driven/called by the application. It may be reading from a local csv
file, from a remote service, from an external database.

Added a test where if the given loan amount exceeds the lenders'
available budge, the api should raise a problem. For the logic test, we
have used fetching the lenders hardcoded from memory, to pass the
behaviour we have defined in our api.