Implemented finding the lowest possible annual interest rate
By finding the lowest possible annual interest rate, what we care about
to give a quick initial quote (not completely accurate as to what the
actual calculations might be when enrolled) is determining the periodic
interest rate. We use that later in the formula to find the periodic
payment; i.e. the payment given the number of months the client wants
to repay their loan.

More info about the formula [here](https://en.wikipedia
.org/wiki/Amortization_calculator#/The_Formula)

Also defined a quick way to compose functions together, so we can reuse
certain small functions; for example rounding up using a given scale.