Preparatory refactor
Now that we are moving towards expanding fetching lenders from a csv
file, we broke down the quote api into a multi-module project. Its core
is the very domain, and the rest modules are going to adapt to its
driving and driven boundaries.

This will also play nicely when we adapt the driving api to be a console
application.