Implemented csv adapter
Added another module responsible to fetch lenders from a csv. To comply
with the api contract, we have added a test.csv file in resources.

We have also added support for testFixtures, in order to share the
contract (base test suite) that is defined in the domain module. So
other technology-specific modules can still run the contract against a
different variation of the api; for example, one that the way of
fetching the lenders is via a local csv file, and not hardcoded in
memory.