Implemented quote adapted to console output
The quote api requires to display its specific format in stdout. This is
another concern, thus another module. To test that this new adapter
complies with the contract, we have opted to consume stdout in an output
stream so that we can parse it and map it the correct domain model.

The driver of this test is a piece of logic which calls the quote api,
exposed in writing into stdout, then maps that stdout into the
respective domain model, and sends it back to comply with what the
contract expects and exercises.

Similarly, if the api was exposed with http concerns, the test driver
would be a http client which calls the server, maps the http-specific
model to the domain, and sends it back to comply with the contract.

Small comment: We defined some `bind` helper functions to partially
apply a function returning another function that requires fewer
arguments. We did that so that it's easier to build the api operation by
partially applying its dependencies beforehand.