Implemented a top level app module
This module is responsible to demo the api in a console manner. Both csv
files are added beforehand in the module's resources folder.

Both files are captured in variables in the main, so to test another
file with a different loan amount, simply point another into the
operation responsible to fetch lenders from a csv file, and modify the
argument (loan amount) in the find quote operation.